move to folders and delete unused files
