+++ service_ui/src/Helpers/consts.ts
