+++ service_ui/src/Helpers/helpers.ts
